+++ hardhat.config.js
@@ -1,6 +1,18 @@
 require("@nomicfoundation/hardhat-toolbox");
-const { getIndexerContract, getStringsLibrary } = require("./helpers/web3");
-const { hardDelete, bypassRevert } = require("./proteins.config");
+const { queryProtein } = require("./helpers/queries");
+const {
+  getIndexerProteinContract,
+  getStringsLibrary,
+  getIndexerSeedContract,
+  getQuerySemiBlastContract,
+  getQueryNaiveContract,
+} = require("./helpers/web3");
+const {
+  hardDelete,
+  bypassRevert,
+  contracts,
+  queryOptions,
+} = require("./proteins.config");
 
 /** @type import('hardhat/config').HardhatUserConfig */
 module.exports = {
@@ -32,15 +44,6 @@ module.exports = {
 
 //Created by Tousuke (zenodeapp - https://github.com/zenodeapp/protein-crud).
 
-/* OWNER CONTRACT */
-task("addAdmin", "Add a new admin.")
-  .addParam("address", "The address.")
-  .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
-
-    await contract.addAdmin(taskArgs.address);
-  });
-
 /* PROTEINS CONTRACT */
 task("deleteProtein", "Delete a protein with the given NFT ID.")
   .addParam("nft", "The NFT ID.")
@@ -52,7 +55,7 @@ task("deleteProtein", "Delete a protein with the given NFT ID.")
   .addOptionalParam("bypass", "Bypass revert.", bypassRevert ? "true" : "false")
   .setAction(async (taskArgs, hre) => {
     const { nft, hard, bypass } = taskArgs;
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerProteinContract(hre);
 
     const result = await contract.deleteProtein(nft, eval(hard), eval(bypass));
     console.log(result);
@@ -62,7 +65,7 @@ task("getProtein", "Returns the protein for the given NFT ID.")
   .addParam("nft", "The NFT ID.")
   .setAction(async (taskArgs, hre) => {
     const { nft } = taskArgs;
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerProteinContract(hre);
 
     const result = await contract.getProtein(nft);
     console.log(result);
@@ -70,7 +73,7 @@ task("getProtein", "Returns the protein for the given NFT ID.")
 
 task("getProteinCount", "How many proteins are included in storage.").setAction(
   async (_, hre) => {
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerProteinContract(hre);
 
     const result = await contract.getProteinCount();
     console.log(result);
@@ -80,11 +83,66 @@ task("getProteinCount", "How many proteins are included in storage.").setAction(
 task("getProteinAtIndex", "Returns the NFT ID at the given index.")
   .addParam("index", "The index value.")
   .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerProteinContract(hre);
 
     const result = await contract.getProteinAtIndex(taskArgs.index);
     console.log(result);
   });
+
+task(
+  "insertSeedAddress",
+  "Inserts a new seed address in the proteins contract."
+)
+  .addParam(
+    "address",
+    "The seed address to add to the protein indexer.",
+    contracts.indexerSeed.address
+  )
+  .setAction(async (taskArgs, hre) => {
+    const contract = await getIndexerProteinContract(hre);
+
+    await contract.insertSeedAddress(taskArgs.address);
+  });
+
+task("updateSeedAddress", "Update the seed address for a given seed size.")
+  .addParam("seedsize", "The seed size.")
+  .addParam("address", "The seed address to add to the protein indexer.")
+  .setAction(async (taskArgs, hre) => {
+    const { seedsize, address } = taskArgs;
+    const contract = await getIndexerProteinContract(hre);
+
+    await contract.updateSeedAddress(seedsize, address);
+  });
+
+task(
+  "deleteSeedAddress",
+  "Delete the stored seed address for a given seed size."
+)
+  .addParam("seedsize", "The seed size.")
+  .setAction(async (taskArgs, hre) => {
+    const contract = await getIndexerProteinContract(hre);
+
+    await contract.deleteSeedAddress(taskArgs.seedsize);
+  });
+
+task("getSeedAddress", "Get the stored seed address for a given seed size.")
+  .addParam("seedsize", "The seed size.")
+  .setAction(async (taskArgs, hre) => {
+    const contract = await getIndexerProteinContract(hre);
+
+    const result = await contract.getSeedAddress(taskArgs.seedsize);
+    console.log(result);
+  });
+
+task(
+  "getIndexerProteinInfo",
+  "Returns information about this indexer."
+).setAction(async (_, hre) => {
+  const contract = await getIndexerProteinContract(hre);
+
+  const result = await contract.getIndexerInfo();
+  console.log(result);
+});
 
 /* SEEDS CONTRACT */
 task("deleteSeed", "Delete the given seed.")
@@ -97,7 +155,7 @@ task("deleteSeed", "Delete the given seed.")
   .addOptionalParam("bypass", "Bypass revert.", bypassRevert ? "true" : "false")
   .setAction(async (taskArgs, hre) => {
     const { nft, hard, bypass } = taskArgs;
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerSeedContract(hre);
 
     const result = await contract.deleteSeed(nft, eval(hard), eval(bypass));
     console.log(result);
@@ -109,7 +167,7 @@ task(
 )
   .addParam("seed", "The seed.")
   .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerSeedContract(hre);
 
     const result = await contract.getSeed(taskArgs.seed);
     console.log(result);
@@ -117,7 +175,7 @@ task(
 
 task("getSeedCount", "How many seeds are included in storage.").setAction(
   async (_, hre) => {
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerSeedContract(hre);
 
     const result = await contract.getSeedCount();
     console.log(result);
@@ -127,11 +185,20 @@ task("getSeedCount", "How many seeds are included in storage.").setAction(
 task("getSeedAtIndex", "Returns the seed at the given index.")
   .addParam("index", "The index value.")
   .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
+    const contract = await getIndexerSeedContract(hre);
 
     const result = await contract.getSeedAtIndex(taskArgs.index);
     console.log(result);
   });
+
+task("getIndexerSeedInfo", "Returns information about this indexer.").setAction(
+  async (_, hre) => {
+    const contract = await getIndexerSeedContract(hre);
+
+    const result = await contract.getIndexerInfo();
+    console.log(result);
+  }
+);
 
 /* STRINGS LIBRARY */
 task("fragmentWord", "Split a word in multiple segments.")
@@ -160,66 +227,112 @@ task("fragmentWord", "Split a word in multiple segments.")
     console.log(result);
   });
 
-// task("seedSize", "Get the seed size.").setAction(async (_, hre) => {
-//   const contract = await getIndexerContract(hre);
-
-//   const result = await contract.seedSize();
-//   console.log(`Seed size is ${result}`);
-// });
-// task("seedStep", "Get the seed step.").setAction(async (_, hre) => {
-//   const contract = await getIndexerContract(hre);
-
-//   const result = await contract.seedStep();
-//   console.log(`Seed step is ${result}`);
-// });
-
-// task("updateSeedSize", "Set a new value for the seed size.")
-//   .addParam("size", "The size.")
-//   .setAction(async (taskArgs, hre) => {
-//     const contract = await getIndexerContract(hre);
-
-//     await contract.updateSeedSize(parseInt(taskArgs.size));
-//   });
-// task("updateSeedStep", "Set a new value for the seed step.")
-//   .addParam("step", "The step.")
-//   .setAction(async (taskArgs, hre) => {
-//     const contract = await getIndexerContract(hre);
-
-//     await contract.updateSeedStep(parseInt(taskArgs.step));
-//   });
-
-/* QUERIES */
-task("naiveQuery", "Get all NFTs matching id or sequence with the given query.")
+/* QUERYING */
+task(
+  "semiBlastQuery",
+  "Get all NFTs matching the query using the semi-blast query protocol."
+)
   .addOptionalParam("sequence", "The sequence query.", "")
-  .addOptionalParam("id", "The PDBID/ACCESSION query.", "")
   .addOptionalParam(
-    "exclusive",
-    "Exclusive would mean that both query restrictions must be true, else one or the other has to be true.",
-    "false"
+    "seedsize",
+    "Seed size to use for querying.",
+    queryOptions.defaultSeedSize.toString()
+  )
+  .addOptionalParam(
+    "limit",
+    "Limit the amount of results given back.",
+    queryOptions.defaultLimit.toString()
+  )
+  .addOptionalParam(
+    "casesensitive",
+    "Case-sensitive lookup.",
+    queryOptions.defaultCaseSensitivity ? "true" : "false"
+  )
+  .addOptionalParam(
+    "format",
+    'How would you like the data to formatted? Choose between "protein", "id", "nft", "sequence" or "ipfs".',
+    queryOptions.defaultFormat
   )
   .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
-    const { id, sequence, exclusive } = taskArgs;
+    const { sequence, seedsize, limit, casesensitive, format } = taskArgs;
 
-    const result = await contract.naiveQuery(id, sequence, eval(exclusive));
-    console.log(result.proteins);
-    console.log(
-      `${result.proteinsFound} results found matching query {sequence: "${sequence}", id: "${id}", exclusive: ${exclusive}}.`
-    );
+    const contract = await getQuerySemiBlastContract(hre);
+    const queryInput = { sequence };
+    const queryOptions = {
+      seedSize: parseInt(seedsize),
+      limit: parseInt(limit),
+      caseSensitive: eval(casesensitive),
+    };
+
+    if (contract) {
+      const result = await queryProtein(
+        contract,
+        queryInput,
+        queryOptions,
+        format
+      );
+
+      queryOptions.format = format;
+
+      console.log(result.data);
+      console.log();
+      console.log(`${result.count} results found matching query:`);
+      console.log(queryInput);
+      console.log(queryOptions);
+    }
   });
 
 task(
-  "semiBlastQuery",
-  "Get all NFTs matching the query using the semi-blast protocol."
+  "naiveQuery",
+  "Get all NFTs matching the query using the naive query protocol."
 )
   .addOptionalParam("sequence", "The sequence query.", "")
+  .addOptionalParam("id", "The PDBID/ACCESSION query.", "")
+  .addOptionalParam(
+    "limit",
+    "Limit the amount of results given back.",
+    queryOptions.defaultLimit.toString()
+  )
+  .addOptionalParam(
+    "casesensitive",
+    "Case-sensitive lookup.",
+    queryOptions.defaultCaseSensitivity ? "true" : "false"
+  )
+  .addOptionalParam(
+    "union",
+    "Union query would combine all individual results into one set. E.g. sequence = AAA and id = 1A, would result in all matches for AAA and all matches for 1A.",
+    queryOptions.defaultUnion ? "true" : "false"
+  )
+  .addOptionalParam(
+    "format",
+    'How would you like the data to formatted? Choose between "protein", "id", "nft", "sequence" or "ipfs".',
+    queryOptions.defaultFormat
+  )
   .setAction(async (taskArgs, hre) => {
-    const contract = await getIndexerContract(hre);
-    const { sequence } = taskArgs;
+    const { id, sequence, limit, union, casesensitive, format } = taskArgs;
 
-    const result = await contract.semiBlastQuery(sequence);
-    console.log(result.proteins);
-    console.log(
-      `${result.proteinsFound} results found matching query {sequence: "${sequence}"}.`
-    );
+    const contract = await getQueryNaiveContract(hre);
+    const queryInput = { id, sequence };
+    const queryOptions = {
+      limit: parseInt(limit),
+      caseSensitive: eval(casesensitive),
+      union: eval(union),
+    };
+
+    if (contract) {
+      const result = await queryProtein(
+        contract,
+        queryInput,
+        queryOptions,
+        format
+      );
+
+      queryOptions.format = format;
+
+      console.log(result.data);
+      console.log();
+      console.log(`${result.count} results found matching query:`);
+      console.log(queryInput);
+      console.log(queryOptions);
+    }
   });