+++ helpers/web3.js
@@ -2,36 +2,72 @@
 
 const { contracts, libraries } = require("../proteins.config");
 
-const getIndexerContract = async (hre) => {
+const getIndexerProteinContract = async (hre) => {
   const contract = await hre.ethers.getContractAt(
-    contracts.indexer.name,
-    contracts.indexer.address
+    contracts.indexerProtein.name,
+    contracts.indexerProtein.address
   );
 
   return contract;
 };
 
-const getQueryContract = async (hre) => {
+const getIndexerSeedContract = async (hre) => {
   const contract = await hre.ethers.getContractAt(
-    contracts.query.name,
-    contracts.query.address
+    contracts.indexerSeed.name,
+    contracts.indexerSeed.address
   );
 
   return contract;
 };
 
-const getIndexerFactory = async (hre, config) => {
+const getQuerySemiBlastContract = async (hre) => {
+  const contract = await hre.ethers.getContractAt(
+    contracts.querySemiBlast.name,
+    contracts.querySemiBlast.address
+  );
+
+  return contract;
+};
+
+const getQueryNaiveContract = async (hre) => {
+  const contract = await hre.ethers.getContractAt(
+    contracts.queryNaive.name,
+    contracts.queryNaive.address
+  );
+
+  return contract;
+};
+
+const getIndexerSeedFactory = async (hre, config) => {
   const Factory = await hre.ethers.getContractFactory(
-    contracts.indexer.name,
+    contracts.indexerSeed.name,
     config
   );
 
   return Factory;
 };
 
-const getQueryFactory = async (hre, config) => {
+const getIndexerProteinFactory = async (hre, config) => {
   const Factory = await hre.ethers.getContractFactory(
-    contracts.query.name,
+    contracts.indexerProtein.name,
+    config
+  );
+
+  return Factory;
+};
+
+const getQuerySemiBlastFactory = async (hre, config) => {
+  const Factory = await hre.ethers.getContractFactory(
+    contracts.querySemiBlast.name,
+    config
+  );
+
+  return Factory;
+};
+
+const getQueryNaiveFactory = async (hre, config) => {
+  const Factory = await hre.ethers.getContractFactory(
+    contracts.queryNaive.name,
     config
   );
 
@@ -47,6 +83,15 @@ const getStringsLibrary = async (hre) => {
   return contract;
 };
 
+const getStringsFactory = async (hre, config) => {
+  const Factory = await hre.ethers.getContractFactory(
+    libraries.strings.name,
+    config
+  );
+
+  return Factory;
+};
+
 const getStructsLibrary = async (hre) => {
   const contract = await hre.ethers.getContractAt(
     libraries.structs.name,
@@ -56,11 +101,26 @@ const getStructsLibrary = async (hre) => {
   return contract;
 };
 
+const getStructsFactory = async (hre, config) => {
+  const Factory = await hre.ethers.getContractFactory(
+    libraries.structs.name,
+    config
+  );
+
+  return Factory;
+};
+
 module.exports = {
-  getIndexerContract,
-  getQueryContract,
-  getIndexerFactory,
-  getQueryFactory,
+  getIndexerProteinContract,
+  getIndexerProteinFactory,
+  getIndexerSeedContract,
+  getIndexerSeedFactory,
+  getQuerySemiBlastContract,
+  getQuerySemiBlastFactory,
+  getQueryNaiveContract,
+  getQueryNaiveFactory,
   getStringsLibrary,
+  getStringsFactory,
   getStructsLibrary,
+  getStructsFactory,
 };