+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 ## BIG DATA ERA
 
+#### 1.2.3 (2022-08-18)
+
+- Naive algorithm now supports \*-wildcard functionality (but works as ?-wildcards, see: https://support.microsoft.com/en-us/office/examples-of-wildcard-characters-939e153f-bd30-47e4-a763-61897c87b3f4)
+- Union querying has been implemented in the naive algorithm (details: https://www.sqlshack.com/sql-union-overview-usage-and-examples/).
+- Refactored and made the algorithms more cohesive.
+- Helper function queryProtein created and refactored the tasks in hardhat.config.js.
+- Format flag implemented, this allows us to change the return values when querying. Available values are: "protein", "nft", "id", "sequence" and "ipfs".
+- Tasks minimized to: `npx hardhat queryNaive` and `npx hardhat querySemiBlast`. Use the new --format flag to change what results you want.
+
+#### 1.2.2 (2022-08-16)
+
+- Separated Naive and Semi-Blast algorithms into two contracts.
+- It's possible now to query IDs (naive approach), but separate from sequences.
+- BREAKING CHANGES: Removed the TempQuery contract and all its scripts.
+- BREAKING CHANGES: (Temporarily) disabled NaiveQuery.
+- BREAKING CHANGES: Temporarily disabled querying words smaller than the seedSize.
+- Case (in)sensitive functionality for Semi-Blast added.
+- Limiting the amount of results functionality added.
+- Return different datasets on every query.
+- Separated all contracts successfully. IndexerProtein, IndexerSeed and Query are now the three main contracts.
+- Protein Pagination idea is somewhat born by this separation of contracts.
+- Separation of contracts now allows for linking multiple seed sizes to one protein contract.
+- Total amount of (detectable) positions can be queried via getIndexerInfo in the IndexerSeed Contract. Detectable positions are the positions that haven't been soft-deleted.
+- Bugfix: querying for proteins that weren't added yet resulted in a crash.
+- IndexerProtein is linked to the IndexerSeed by storing the seed's contract address in the protein contract.
+
 #### 1.2.1 (2022-08-11)
 
 - Changed the delete mechanism for seed positions in CrudSeed.sol.